--- backend/src/main/java/kr/co/nogibackend/application/notion/NotionPageFacade.java
@@ -1,15 +1,0 @@
-package kr.co.nogibackend.application.notion;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class NotionPageFacade {
-
-  // 작성완료 페이지 처리
-  public void processCompletedPages() {
-
-  }
-
-}
