+++ backend/src/main/java/kr/co/nogibackend/domain/notion/dto/info/NotionBlockConversionInfo.java
@@ -1,11 +1,11 @@
 package kr.co.nogibackend.domain.notion.dto.info;
 
 import java.util.List;
-import kr.co.nogibackend.domain.notion.dto.result.NotionStartTILResult;
+import kr.co.nogibackend.domain.notion.dto.result.CompletedPageMarkdownResult;
 
 public record NotionBlockConversionInfo(
     String content,
-    List<NotionStartTILResult.ImageOfNotionBlock> images
+    List<CompletedPageMarkdownResult.ImageOfNotionBlock> images
 ) {
 
 }
